output is now a part of the Pipeline struct
